old_bot/requirements.txt
python-telegram-bot[callback-data]
pydantic
aiohttp
gql
thefuzz
